requirements.txt
alembic==1.0.5
Click==7.0
Flask==1.0.2
Flask-Admin==1.5.3
Flask-Login==0.4.1
Flask-Migrate==2.3.1
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
gunicorn==19.9.0
itsdangerous==1.1.0
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.1.0
Pillow==5.4.0
python-dateutil==2.7.5
python-editor==1.0.3
six==1.12.0
SQLAlchemy==1.2.15
Werkzeug==0.14.1
WTForms==2.2.1
